third_party/blink/web_tests/http/tests/devtools/components/throttler-expected.txt

This test verifies throttler behavior.


Running: testSimpleSchedule
Throttler is in IDLE state (doesn't have any timers set up)
SCHEDULED: 'operation #1' asSoonAsPossible: Default
SCHEDULED: 'operation #2' asSoonAsPossible: undefined
Throttler is in TIMEOUT state. Scheduled timers timeouts: [1989]
Process 'operation #2' STARTED.
Process 'operation #2' FINISHED.

Running: testAsSoonAsPossibleOverrideTimeout
Throttler is in IDLE state (doesn't have any timers set up)
SCHEDULED: 'operation #1' asSoonAsPossible: undefined
SCHEDULED: 'operation #2' asSoonAsPossible: AsSoonAsPossible
Throttler is in TIMEOUT state. Scheduled timers timeouts: [0]
Process 'operation #2' STARTED.
Process 'operation #2' FINISHED.

Running: testAlwaysExecuteLastScheduled
Throttler is in IDLE state (doesn't have any timers set up)
SCHEDULED: 'operation #0' asSoonAsPossible: AsSoonAsPossible
SCHEDULED: 'operation #1' asSoonAsPossible: Default
SCHEDULED: 'operation #2' asSoonAsPossible: AsSoonAsPossible
SCHEDULED: 'operation #3' asSoonAsPossible: Default
Throttler is in TIMEOUT state. Scheduled timers timeouts: [0]
Process 'operation #3' STARTED.
Process 'operation #3' FINISHED.

Running: testSimpleScheduleDuringProcess
Throttler is in IDLE state (doesn't have any timers set up)
SCHEDULED: 'long operation' asSoonAsPossible: undefined
Throttler is in TIMEOUT state. Scheduled timers timeouts: [1989]
Process 'long operation' STARTED.
Throttler is in IDLE state (doesn't have any timers set up)
SCHEDULED: 'operation #1' asSoonAsPossible: Default
SCHEDULED: 'operation #2' asSoonAsPossible: undefined
Process 'long operation' FINISHED.
Throttler is in TIMEOUT state. Scheduled timers timeouts: [1989]
Process 'operation #2' STARTED.
Process 'operation #2' FINISHED.

Running: testAsSoonAsPossibleOverrideDuringProcess
Throttler is in IDLE state (doesn't have any timers set up)
SCHEDULED: 'long operation' asSoonAsPossible: undefined
Throttler is in TIMEOUT state. Scheduled timers timeouts: [1989]
Process 'long operation' STARTED.
Throttler is in IDLE state (doesn't have any timers set up)
SCHEDULED: 'operation #1' asSoonAsPossible: undefined
SCHEDULED: 'operation #2' asSoonAsPossible: AsSoonAsPossible
Process 'long operation' FINISHED.
Throttler is in TIMEOUT state. Scheduled timers timeouts: [0]
Process 'operation #2' STARTED.
Process 'operation #2' FINISHED.

Running: testAlwaysExecuteLastScheduledDuringProcess
Throttler is in IDLE state (doesn't have any timers set up)
SCHEDULED: 'long operation' asSoonAsPossible: undefined
Throttler is in TIMEOUT state. Scheduled timers timeouts: [1989]
Process 'long operation' STARTED.
Throttler is in IDLE state (doesn't have any timers set up)
SCHEDULED: 'operation #0' asSoonAsPossible: AsSoonAsPossible
SCHEDULED: 'operation #1' asSoonAsPossible: Default
SCHEDULED: 'operation #2' asSoonAsPossible: AsSoonAsPossible
SCHEDULED: 'operation #3' asSoonAsPossible: Default
Process 'long operation' FINISHED.
Throttler is in TIMEOUT state. Scheduled timers timeouts: [0]
Process 'operation #3' STARTED.
Process 'operation #3' FINISHED.

Running: testScheduleFromProcess
Throttler is in IDLE state (doesn't have any timers set up)
SCHEDULED: 'operation #1' asSoonAsPossible: undefined
Throttler is in TIMEOUT state. Scheduled timers timeouts: [1989]
Process 'operation #1' STARTED.
SCHEDULED: 'operation #2' asSoonAsPossible: Default
Process 'operation #1' FINISHED.
Throttler is in TIMEOUT state. Scheduled timers timeouts: [1989]
Process 'operation #2' STARTED.
Process 'operation #2' FINISHED.

Running: testExceptionFromProcess
SCHEDULED: 'operation #1' asSoonAsPossible: undefined
Process 'operation #1' STARTED.
error: Error: Exception during process execution.
Throttler is in IDLE state (doesn't have any timers set up)

Running: testPromise
SCHEDULED: 'operation #1' asSoonAsPossible: undefined
Process 'operation #1' STARTED.
Process 'operation #1' FINISHED.
The promise resolved.
